Allow the debugger to set an explicit libGL path
When trying to resolve the system OpenGL module path we do not respect
/etc/ld.so.conf and friends. This prevents detection of OpenGL libraries
in unexpected locations (eg. nVidia modules are located at
/usr/lib64/opengl/nvidia/lib under Gentoo Linux). As CodeXL may
indirectly pull in the 'proper' system OpenGL module via Qt, adds it's
own spies library, and explicitly loads a module from a more typical
path we might end up with multiple OpenGL libraries in the one process.

We should query the system library loader to resolve the path to
libGL.so.1 on our behalf; it is not possible to do this ourselves
reliably due to the opaque format of /etc/ld.so.cache

The two methods used for library interposition on Linux-like systems
(LD_PRELOAD and LD_LIBRARY_PATH) both put the spies library ahead of the
system OpenGL implementation. This makes it difficult for the debugged
client (and hence the spies library) to reliably detect the library that
would have been loaded on an unpertubed system. Instead, we set/get a
well known environment variable (SU_SYSTEM_OPENGL_MODULE_PATH) from the
debugging server with the unpurturbed value.

This technique isn't required for OpenCL or Vulkan due to the explicit
paths in their ICD configuration.