Update the code that generates occupancy HTML files.
It was using a vector to see if a file had already been generated
for a given occupancy index.  The problem is that is wasn't checking
the session so if occupancy for index 271 was generated in session A,
it would fail to generated occupancy for index 271 in Session B even
though it would have needed to regenerate it. To fix this, we now use
a map from session id to vector of generated indexes.

Also, don't try to use the debug version of rcprof in a debug build, since
you normally won't have a debug build of the profiler.